# Conflicts:
#	c3-employee-portal/src/app/member-search-result/member-search-result.component.html
#	c3-employee-portal/src/app/member-search-result/member-search-result.component.ts
#	c3-employee-portal/src/app/membersearch/membersearch.component.ts
#	c3-employee-portal/src/app/primaryhousehold/primaryhousehold.component.ts

+++ c3-checkin/src/app/services/services.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit} from '@angular/core';
+import {AfterViewInit, Component, OnInit} from '@angular/core';
 import {Service} from '../../models/service';
 import {Member} from '../../models/member';
 import {ServicesByMember} from '../../models/servicesByMember';
@@ -15,10 +15,10 @@ export class ServicesComponent implements OnInit {
   whoNeeds = [];
   selected = [];
   services = [];
+  alreadyLoaded: any;
 
   constructor(private apiService: ApiService, private router: Router) {
     this.services = [];
-
     this.apiService.getServices()
       .subscribe((data) => {
         let i;
@@ -27,13 +27,18 @@ export class ServicesComponent implements OnInit {
           this.services.push(new Service(data[i].service_id, data[i].name));
         }
       });
-
-    this.whoNeeds = JSON.parse(sessionStorage.getItem('whoNeeds'));
-    this.selectedMember = this.whoNeeds.pop();
-    sessionStorage.setItem('whoNeeds', JSON.stringify(this.whoNeeds, null, 4));
   }
 
   ngOnInit() {
+    if (!this.alreadyLoaded) {
+      this.whoNeeds = JSON.parse(sessionStorage.getItem('whoNeeds'));
+      this.selectedMember = this.whoNeeds.pop();
+      sessionStorage.setItem('whoNeeds', JSON.stringify(this.whoNeeds, null, 4));
+
+      this.selected = [];
+      this.alreadyLoaded = true;
+    }
+
     if (this.selectedMember == null) {
       this.router.navigate(['login']);
     }
@@ -94,8 +99,23 @@ export class ServicesComponent implements OnInit {
       sessionStorage.setItem('servicesByMember', JSON.stringify(servicesByMember, null, 4));
 
       if (this.whoNeeds.length > 0) {
-        location.reload();
+        this.alreadyLoaded = false;
+        this.ngOnInit();
       } else {
+        // [
+        //   {
+        //     "member_id": "3fb6798c-c1cd-47d8-a123-49812f204cba",
+        //     "service_id": "3fb6798c-c1cd-47d8-a123-49812f204cba",
+        //     "created_date": "20191020T013132"
+        //   }
+        // ]
+
+        // this.apiService.submitServicesByMember().subscribe((data) => {
+        //   console.log(data);
+        //
+        //   this.router.navigate(['thanks']);
+        // });
+
         this.router.navigate(['thanks']);
       }
     } else {
